More cosmetics (unit tests)
Split PlayResultSpec and EventSpec into CheckerSpec
